Init helper func to return normalized data.
